Add python single threaded cached script
Add an python script, which runs single threaded, but uses a cached
approach to find prime numbers.
By cached approach, it means, when generating a list of prime numbers,
it uses previous prime numbers, to check next number is divisible by, instead of all number below the next number.

This approach works, because every number which is not a prime number,
is a composite of some prime number smaller than it.